Moved all into plugin.py and added synchronous/asynchronous guards
